+++ vpi/sys_readmem.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1999-2009 Stephen Williams (steve@icarus.com)
+ * Copyright (c) 1999-2010 Stephen Williams (steve@icarus.com)
  *
  *    This source code is free software; you can redistribute it
  *    and/or modify it in source code form under the terms of the GNU
@@ -450,7 +450,7 @@ static PLI_INT32 sys_readmempath_calltf(PLI_BYTE8*name)
 	 */
       len = strlen(val.value.str);
       for (idx = 0; idx < len; idx++) {
-	    if (! isprint(val.value.str[idx])) {
+	    if (! isprint((int)val.value.str[idx])) {
 		  char msg [64];
 		  char *esc_path = as_escaped(val.value.str);
 		  snprintf(msg, 64, "WARNING: %s:%d:",
