+++ vpi/vcd_priv.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003-2009 Stephen Williams (steve@icarus.com)
+ * Copyright (c) 2003-2010 Stephen Williams (steve@icarus.com)
  *
  *    This source code is free software; you can redistribute it
  *    and/or modify it in source code form under the terms of the GNU
@@ -36,11 +36,11 @@ int is_escaped_id(const char *name)
 
       assert(name);
         /* The first digit must be alpha or '_' to be a normal id. */
-      if (isalpha(name[0]) || name[0] == '_') {
+      if (isalpha((int)name[0]) || name[0] == '_') {
 	    for (lp=1; name[lp] != '\0'; lp++) {
 		    /* If this digit is not alpha-numeric or '_' we have
 		     * an escaped identifier. */
-		  if (!(isalnum(name[lp]) || name[lp] == '_')) {
+		  if (!(isalnum((int)name[lp]) || name[lp] == '_')) {
 		        return 1;
 		  }
 	    }
